Move lake.ts to src/models
