+++ apps/zui/src/core/main/main-object.ts
@@ -24,7 +24,7 @@ import {getAppMeta, AppMeta} from "../../electron/meta"
 import {createMainStore} from "../../js/state/stores/create-main-store"
 import {AppDispatch, State} from "../../js/state/types"
 import {PathName, getPath} from "../../js/api/core/get-path"
-import {Lake} from "src/js/models/lake"
+import {Lake} from "src/models/lake"
 import {getAuthToken} from "../../js/api/core/get-zealot"
 import {Abortables} from "src/modules/abortables"
 import * as zui from "src/zui"
