+++ apps/zui/src/domain/lake/handlers.ts
@@ -1,6 +1,6 @@
 import {Client} from "@brimdata/zed-js"
 import {createHandler} from "src/core/handlers"
-import {Lake} from "src/js/models/lake"
+import {Lake} from "src/models/lake"
 import Current from "src/js/state/Current"
 import {validateToken} from "src/core/auth0/utils"
 import {getAuthCredentials} from "src/js/flows/lake/getAuthCredentials"
