+++ apps/zui/src/js/api/zui-api.ts
@@ -13,7 +13,7 @@ import {EditorApi} from "./editor/editor-api"
 import {NoticeApi} from "./notice/notice-api"
 import {UrlApi} from "./url/url-api"
 import {TableViewApi} from "src/zui-kit"
-import {Lake} from "../models/lake"
+import {Lake} from "../../models/lake"
 
 // This is deprecated in favor of domain handlers and operations
 export default class ZuiApi {
