+++ apps/zui/src/js/components/LakeModals/LakeForm.tsx
@@ -9,7 +9,7 @@ import {LakeAttrs} from "../../state/Lakes/types"
 import useCallbackRef from "../hooks/useCallbackRef"
 import useEventListener from "../hooks/useEventListener"
 import {isDefaultLake} from "../../initializers/initLakeParams"
-import {Lake} from "src/js/models/lake"
+import {Lake} from "src/models/lake"
 import {ErrorWell} from "src/components/error-well"
 
 type Props = {
