+++ apps/zui/src/js/components/Login.tsx
@@ -8,7 +8,7 @@ import MacSpinner from "./MacSpinner"
 import {isString} from "lodash"
 import {updateStatus} from "../flows/lake/update-status"
 import {login} from "../flows/lake/login"
-import {Lake} from "../models/lake"
+import {Lake} from "../../models/lake"
 
 const PageWrap = styled.div`
   width: 100%;
