+++ apps/zui/src/js/state/Current/selectors.ts
@@ -15,7 +15,7 @@ import {ActiveQuery} from "src/models/active-query"
 import SessionQueries from "../SessionQueries"
 import memoizeOne from "memoize-one"
 import {entitiesToArray} from "../utils"
-import {Lake} from "src/js/models/lake"
+import {Lake} from "src/models/lake"
 import {defaultLake} from "src/js/initializers/initLakeParams"
 import {getActive} from "../Tabs/selectors"
 import QueryInfo from "../QueryInfo"
